Time: 76 ms (82.24%) | Memory: 56.1 MB (44.15%) - LeetSync

+++ 398-random-pick-index/random-pick-index.java
@@ -12,7 +12,7 @@ class Solution {
     
     public int pick(int target) {
         ArrayList<Integer> arr=hash.get(target);
-        return arr.get((int)(Math.random()*(double)arr.size()));
+        return arr.get((int)(Math.random()*arr.size()));
     }
 }
